fixed creator delete issue. Cannot delete creator unless any items with that creator have been deleted first

+++ Library-Backend/src/main/java/ca/mcgill/ecse321/library/service/CreatorService.java
@@ -85,6 +85,10 @@ public class CreatorService {
 		if (creator == null) {
 			throw new IllegalArgumentException("Creator does not exist.");
 		}
+		List<Item> items = getItemsByCreator(creatorId);
+		if (items.size() > 0) {
+			throw new IllegalArgumentException("Cannot delete creator that has items.");
+		}
 		creatorRepository.delete(creator);
 		return creator;
